Release 1.0.11 * Optimized graph widget relative to axis objects, logarithmic scaling   is now implemented in SSE2 and is about 6 times faster than native implementation. * Refactored widget rendering subsystem: FPS is lowered and stabilized around 20 FPS. * Implemented additional compression of built-in XML resources for the UI, now   XML resources obtain about 3 times lower space in binaries. * Highly optimized rendering of Gtk2Graph and Gtk2File widgets that caused excessive CPU load:   the CPU utilization is now about 7 times lower.

+++ src/ui/graphics.cpp
@@ -69,7 +69,7 @@ namespace lsp
     {
         float dx = x1 - x2;
         float dy = y1 - y2;
-        return sqrt(dx*dx + dy*dy);
+        return sqrtf(dx*dx + dy*dy);
     }
 
     float scalar_product2d(float x1, float y1, float x2, float y2)
@@ -90,7 +90,7 @@ namespace lsp
     {
         float dx = x - x0;
         float dy = y - y0;
-        float l = sqrt(dx*dx + dy*dy);
+        float l = sqrtf(dx*dx + dy*dy);
         if (l == 0.0)
             return 0;
